ci: Pin versions of GitHub Actions to full commit hash (#120)
## What?

This PR pins versions of GitHub Actions to full commit hash by pinact.
In general, this PR doesn't change the behavior of the workflows, so you
can merge this safely.

## Why?


https://docs.github.com/en/actions/security-for-github-actions/security-guides/security-hardening-for-github-actions#using-third-party-actions

> Pin actions to a full length commit SHA
> Pinning an action to a full length commit SHA is currently the only
way to use an action as an immutable release.
> Pinning to a particular SHA helps mitigate the risk of a bad actor
adding a backdoor to the action's repository, as they would need to
generate a SHA-1 collision for a valid Git object payload.
> When selecting a SHA, you should verify it is from the action's
repository and not a repository fork.

## How was this pull request created?

This pull request was created by
[pinact](https://github.com/suzuki-shunsuke/pinact) and
[multi-gitter](https://github.com/lindell/multi-gitter).

## Due Date

Please merge this pull request by 202X-XX-XX.
If it's difficult, please ask at the Slack channel #sre.

## Contact

If you have any question, please ask at the Slack channel #sre.